integrated team modules

+++ Bookstore/src/main/java/com/cg/bookstore/entities/QueryResponseDTO.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class QueryResponseDTO {
 
 	 private List<CustomerInformation> list;
-	 private int totalNoOfPages;
+	 private long totalNoOfPages;
 	 private int currentPageNumber;
 	public List<CustomerInformation> getList() {
 		return list;
@@ -13,10 +13,10 @@ public class QueryResponseDTO {
 	public void setList(List<CustomerInformation> list) {
 		this.list = list;
 	}
-	public int getTotalNoOfPages() {
+	public long getTotalNoOfPages() {
 		return totalNoOfPages;
 	}
-	public void setTotalNoOfPages(int totalNoOfPages) {
+	public void setTotalNoOfPages(long totalNoOfPages) {
 		this.totalNoOfPages = totalNoOfPages;
 	}
 	public int getCurrentPageNumber() {
